% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admml1_grad.R
\name{admml1_grad}
\alias{admml1_grad}
\title{admml1_grad  function}
\usage{
admml1_grad(theta, x, alpha, u, G, Dk, rho)
}
\arguments{
\item{theta}{vector}

\item{x}{data}

\item{alpha}{ADMM parameter}

\item{u}{vector}

\item{G}{matrix}

\item{Dk}{matrix}

\item{rho}{scalar}
}
\description{
This function allows to compute admm gradient
}
