% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admml2_grad.R
\name{admml2_grad}
\alias{admml2_grad}
\title{admml2_grad  function}
\usage{
admml2_grad(theta, x, b, G, D, L, alpha)
}
\arguments{
\item{theta}{vector}

\item{x}{data}

\item{G}{matrix}

\item{alpha}{ADMM parameter}

\item{u}{vector}

\item{Dk}{matrix}

\item{rho}{scalar}
}
\description{
This function allows to compute admm gradient for L2 deconvol
}
