% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admml2_val.R
\name{admml2_val}
\alias{admml2_val}
\title{admml2_val  function}
\usage{
admml2_val(theta, x, b, G, D, L, alpha)
}
\arguments{
\item{theta}{vector}

\item{x}{data}

\item{G}{matrix}

\item{alpha}{ADMM parameter}

\item{u}{vector}

\item{Dk}{matrix}

\item{rho}{scalar}
}
\description{
This function allows to compute admm objective for L2 deconvol
}
