% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnormix.R
\name{dnormix}
\alias{dnormix}
\title{dnormix function}
\usage{
dnormix(x, weights, mu, tau2)
}
\arguments{
\item{x}{locations where to evaluate the pdf}

\item{weights}{weights on each mixture component}

\item{tau2}{variances of mixture components}

\item{means}{centers of mixture components}
}
\description{
This function allows to evaluate the pdf of a mixture of normals
}
