% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridprep.R
\name{gridprep}
\alias{gridprep}
\title{gridprep function}
\usage{
gridprep(y, gridsize = 250)
}
\arguments{
\item{y}{raw observations for deconvolution}

\item{gridsize}{this is the number of bins}
}
\description{
Preparing data for calculations on grid
}
