% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_deconvolution.R
\name{L1_deconvolution}
\alias{L1_deconvolution}
\title{L1  deconconvolution function}
\usage{
L1_deconvolution(y, d)
}
\arguments{
\item{y}{: a vector containing the raw observations.}

\item{d}{: number of bins}
}
\value{
loc : location of the centers of bins where the density is estimated

f_hat : estimated mixing density

y_hat : marginal density
}
\description{
This function allows to perform L1 deconvolution with model selection as in the paper "A deconvolution path for mixtures"
}
