% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmltheta.R
\name{logmltheta}
\alias{logmltheta}
\title{logmltheta function}
\usage{
logmltheta(theta, counts, G)
}
\arguments{
\item{theta}{vector}

\item{counts}{vector}

\item{G}{matrix}
}
\description{
logarithm of log-likehood
}
