% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmlthetagrad.R
\name{logmlthetagrad}
\alias{logmlthetagrad}
\title{logmlthetagrad function}
\usage{
logmlthetagrad(theta, counts, G)
}
\arguments{
\item{theta}{vector}

\item{counts}{vector}

\item{G}{matrix}
}
\description{
logarithm of log-likehood
}
