% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_grad.R
\name{obj_grad}
\alias{obj_grad}
\title{obj_grad  function}
\usage{
obj_grad(theta, x, G, L, lambda)
}
\arguments{
\item{theta}{vector}

\item{x}{vector}

\item{G}{matrix}

\item{L}{matrix}

\item{lambda}{scalar}
}
\description{
compute gradient of objective in L2 deconvolution
}
