% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_val.R
\name{obj_val}
\alias{obj_val}
\title{obj_val function}
\usage{
obj_val(theta, x, G, L, lambda)
}
\arguments{
\item{theta}{vector}

\item{x}{vector}

\item{G}{matrix}

\item{L}{matrix}

\item{lambda}{scalar}
}
\description{
objective function in L2 deconvolution
}
