% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormix.R
\name{rnormix}
\alias{rnormix}
\title{rnormix function}
\usage{
rnormix(n, weights, means, tau2)
}
\arguments{
\item{n}{number of samples}

\item{weights}{weights on each mixture component}

\item{means}{centers of mixture components}

\item{tau2}{variances of mixture components}
}
\description{
This function allows to generate sample from a mixture of normals
}
