% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\name{trapezoid}
\alias{trapezoid}
\title{trapezoid function}
\usage{
trapezoid(grid, f)
}
\arguments{
\item{grid}{locations of points}

\item{f}{function values}
}
\description{
trapezoind rule for integrals
}
